[IMP] ni_community_care: move cron job to ni_community_care
