+++ ni_community_care/data/ni_service_event_approval_cron.xml
@@ -1,0 +1,26 @@
+<odoo>
+    <record id="ir_cron_create_my_model" model="ir.cron">
+        <field name="name">Auto Create Monthly Approval</field>
+        <field name="model_id" ref="ni_community_care.model_ni_service_event_approval" />
+        <field name="state">code</field>
+        <field name="code">model.create_record_from_cron()</field>
+        <field name="active" eval="True" />
+        <field name="interval_number">1</field>
+        <field name="interval_type">months</field>
+        <field name="numbercall">-1</field>
+        <!-- ตั้งให้เริ่มรันวันที่ 1 ของเดือนถัดไป เวลา 02:00 -->
+        <field
+            name="nextcall"
+            eval="(DateTime.now().replace(day=1, hour=2, minute=0, second=0) + relativedelta(months=1)).strftime('%Y-%m-%d %H:%M:%S')"
+        />
+    </record>
+    <!--     <record id="ir_cron_generate_approval_reports" model="ir.cron">-->
+    <!--        <field name="name">Generate Event Approval Reports</field>-->
+    <!--        <field name="model_id" ref="ni_community_care.model_ni_service_event_approval"/>-->
+    <!--        <field name="state">code</field>-->
+    <!--        <field name="code">model._cron_generate_reports()</field>-->
+    <!--        <field name="interval_number">1</field>-->
+    <!--        <field name="interval_type">months</field>-->
+    <!--        <field name="active">True</field>-->
+    <!--    </record>-->
+</odoo>
